chore(mgmt-ui): add a minimalist onboarding home page
